+++ hw/xfree86/dri/dri.c
@@ -42,11 +42,14 @@ SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #include <X11/X.h>
 #include <X11/Xfuncproto.h>
 #include <X11/Xproto.h>
+#include <X11/dri/xf86driproto.h>
 
 #include "dix/dix_priv.h"
 #include "dix/screen_hooks_priv.h"
 #include "dix/screenint_priv.h"
+#include "include/dristruct.h"
 #include "include/extinit.h"
+#include "include/sarea.h"
 
 #include "xf86.h"
 #include "xf86drm.h"
@@ -57,12 +60,9 @@ SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #include "scrnintstr.h"
 #include "windowstr.h"
 #include "servermd.h"
-#include <X11/dri/xf86driproto.h>
 #include "swaprep.h"
 #include "xf86str.h"
 #include "dri_priv.h"
-#include "sarea.h"
-#include "dristruct.h"
 #include "mi.h"
 #include "mipointer.h"
 #include "xf86_os_support.h"
