+++ hw/xfree86/dri/xf86dri.c
@@ -37,10 +37,13 @@ SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #include <string.h>
 #include <X11/X.h>
 #include <X11/Xproto.h>
+#include <X11/dri/xf86driproto.h>
 
 #include "dix/dix_priv.h"
 #include "dix/request_priv.h"
 #include "dix/screenint_priv.h"
+#include "include/dristruct.h"
+#include "include/sarea.h"
 
 #include "xf86.h"
 #include "misc.h"
@@ -49,12 +52,9 @@ SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #include "cursorstr.h"
 #include "scrnintstr.h"
 #include "servermd.h"
-#include <X11/dri/xf86driproto.h>
 #include "swaprep.h"
 #include "xf86str.h"
 #include "dri_priv.h"
-#include "sarea.h"
-#include "dristruct.h"
 #include "xf86drm.h"
 #include "protocol-versions.h"
 #include "xf86Extensions.h"
