+++ hw/xquartz/xpr/dri.c
@@ -58,7 +58,7 @@
 #define _APPLEDRI_SERVER_
 #include "appledristr.h"
 #include "swaprep.h"
-#include "dri.h"
+#include "xpr_dri.h"
 #include "dristruct.h"
 #include "mi.h"
 #include "mipointer.h"
