+++ include/meson.build
@@ -429,8 +429,10 @@ if build_xorg
             'dixfont.h',
             'dixfontstr.h',
             'dixstruct.h',
+            'dri.h',
             'dri2.h',
             'dri3.h',
+            'dristruct.h',
             'events.h',
             'exa.h',
             'exevents.h',
@@ -488,6 +490,7 @@ if build_xorg
             'regionstr.h',
             'resource.h',
             'rgb.h',
+            'sarea.h',
             'screenint.h',
             'scrnintstr.h',
             'servermd.h',
