+++ include/sarea.h
